test(gateway): extract relay tests
